daemon: Reload Image Store on a HUP signal
Using an ostree repo for container images storing and transferring
implies updating of images bypassing of the docker daemon API.
If container images are updated by bypassing the docker daemon API (e.g.
`docker pull` or `docker load` or `docker build`) then a running docker
daemon doesn't see such the updated images. The daemon has to be
restarted in order to make it reload image updates since by default
it reads images' and their layers' data from the image and layer store
on a file system to RAM only at the daemon startup.
The given change extends the existing configuration reloading
functionality, that is triggered on receiving SIGHUP, with reloading
of images' and layers' data from a filesystem.

Signed-off-by: Mike Sul <mike.sul@foundries.io>